pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-support_ops_triage_env"
version = "0.1.0"
description = "Real-world support operations triage environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.1",
    "openai>=1.44.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m support_ops_triage_env.server.app
server = "support_ops_triage_env.server.app:main"
baseline = "support_ops_triage_env.scripts.run_baseline:main"
validate_submission = "support_ops_triage_env.scripts.pre_submission_validate:main"

[tool.setuptools]
include-package-data = true
packages = [
    "support_ops_triage_env",
    "support_ops_triage_env.server",
    "support_ops_triage_env.scripts",
]

[tool.setuptools.package-dir]
"support_ops_triage_env" = "."
"support_ops_triage_env.server" = "server"
"support_ops_triage_env.scripts" = "scripts"